requirements-min.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
fastapi==0.115.8
google-auth==2.45.0
google-genai==1.55.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.45.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.30.6
websockets==15.0.1
wheel==0.45.1
